test(portability): split the guard tests at the synthetic/live seam
Review caught a taste ratchet regression this branch introduced and the author's
own verification had declared unrun: the new test file reached 513 lines, taking
the tracked tree to 603 against a baseline of 602. Attribution was proved rather
than inferred, by removing the file from the index and watching the ratchet
return OK at 602.

Raising the baseline would defeat the gate, whose message is explicit that new
error-severity violations cannot merge. The file already held two kinds of test:
synthetic tmp_path exercises of the checker, and assertions that read the live
repository. Splitting on that seam gives 406 and 144 lines.

48 tests pass across both files, taste-lints reports 0 errors on each, and the
count ratchet test passes.
